Rename and move logprep.output.dummy_output to logprep.connector.dummy.output
