+++ tests/unit/connector/test_dummy_output.py
@@ -1,6 +1,6 @@
 from pytest import raises, fail
 
-from logprep.output.dummy_output import DummyOutput
+from logprep.connector.dummy.output import DummyOutput
 from logprep.output.output import FatalOutputError
