+++ tests/unit/framework/test_pipeline.py
@@ -26,7 +26,7 @@ from logprep.abc.input import (
     CriticalInputError,
 )
 from logprep.metrics.metric import MetricTargets
-from logprep.output.dummy_output import DummyOutput
+from logprep.connector.dummy.output import DummyOutput
 from logprep.output.output import FatalOutputError, WarningOutputError, CriticalOutputError
 from logprep.processor.base.exceptions import ProcessingWarning
 from logprep.processor.delete.processor import Delete
@@ -167,7 +167,7 @@ class TestPipeline(ConfigurationForTests):
         self.pipeline.run()
         mock_setup.assert_called()
 
-    @mock.patch("logprep.output.dummy_output.DummyOutput.setup")
+    @mock.patch("logprep.connector.dummy.output.DummyOutput.setup")
     def test_setup_calls_setup_on_output(self, mock_setup, _):
         self.pipeline.run()
         mock_setup.assert_called()
@@ -177,7 +177,7 @@ class TestPipeline(ConfigurationForTests):
         self.pipeline.run()
         mock_shut_down.assert_called()
 
-    @mock.patch("logprep.output.dummy_output.DummyOutput.shut_down")
+    @mock.patch("logprep.connector.dummy.output.DummyOutput.shut_down")
     def test_shut_down_calls_shut_down_on_output(self, mock_shut_down, _):
         self.pipeline.run()
         mock_shut_down.assert_called()
@@ -227,7 +227,7 @@ class TestPipeline(ConfigurationForTests):
         assert len(self.pipeline._output.failed_events) == 1
 
     @mock.patch("logging.Logger.error")
-    @mock.patch("logprep.output.dummy_output.DummyOutput.store")
+    @mock.patch("logprep.connector.dummy.output.DummyOutput.store")
     @mock.patch("logprep.connector.dummy.input.DummyInput.get_next")
     def test_critical_output_error_is_logged_and_stored_as_failed(
         self, mock_get_next, mock_store, mock_error, _
@@ -265,7 +265,7 @@ class TestPipeline(ConfigurationForTests):
         assert len(self.pipeline._output.events) == 2
 
     @mock.patch("logging.Logger.warning")
-    @mock.patch("logprep.output.dummy_output.DummyOutput.store")
+    @mock.patch("logprep.connector.dummy.output.DummyOutput.store")
     @mock.patch("logprep.connector.dummy.input.DummyInput.get_next")
     def test_output_warning_error_is_logged_but_processing_continues(
         self, mock_get_next, mock_store, mock_warning, _
@@ -368,7 +368,7 @@ class TestPipeline(ConfigurationForTests):
         ), "error message is logged"
 
     @mock.patch("logprep.connector.dummy.input.DummyInput.get_next", return_value={"mock": "event"})
-    @mock.patch("logprep.output.dummy_output.DummyOutput.store")
+    @mock.patch("logprep.connector.dummy.output.DummyOutput.store")
     @mock.patch("logging.Logger.error")
     def test_critical_output_error_is_logged(self, mock_error, mock_store, _, __):
         def raise_critical(args):
@@ -385,7 +385,7 @@ class TestPipeline(ConfigurationForTests):
         ), "error message is logged"
 
     @mock.patch("logprep.connector.dummy.input.DummyInput.get_next", return_value={"mock": "event"})
-    @mock.patch("logprep.output.dummy_output.DummyOutput.store")
+    @mock.patch("logprep.connector.dummy.output.DummyOutput.store")
     @mock.patch("logging.Logger.warning")
     def test_warning_output_error_is_logged(self, mock_warning, mock_store, _, __):
         def raise_warning(args):
@@ -415,7 +415,7 @@ class TestPipeline(ConfigurationForTests):
 
     @mock.patch("logprep.connector.dummy.input.DummyInput.get_next", return_value={"mock": "event"})
     @mock.patch("logprep.framework.pipeline.Pipeline._shut_down")
-    @mock.patch("logprep.output.dummy_output.DummyOutput.store")
+    @mock.patch("logprep.connector.dummy.output.DummyOutput.store")
     @mock.patch("logging.Logger.error")
     def test_processor_fatal_output_error_is_logged_pipeline_is_rebuilt(
         self, mock_error, mock_store, mock_shut_down, _, __
@@ -427,7 +427,7 @@ class TestPipeline(ConfigurationForTests):
         assert "Output dummy failed:" in mock_error.call_args[0][0], "error message is logged"
         mock_shut_down.assert_called()
 
-    @mock.patch("logprep.output.dummy_output.DummyOutput.store_custom")
+    @mock.patch("logprep.connector.dummy.output.DummyOutput.store_custom")
     @mock.patch("logprep.connector.dummy.input.DummyInput.get_next", return_value={"mock": "event"})
     def test_extra_dat_tuple_is_passed_to_store_custom(self, mock_get_next, mock_store_custom, _):
         self.pipeline._setup()
@@ -440,7 +440,7 @@ class TestPipeline(ConfigurationForTests):
         mock_store_custom.call_count = 1
         mock_store_custom.assert_called_with({"foo": "bar"}, "target")
 
-    @mock.patch("logprep.output.dummy_output.DummyOutput.store_custom")
+    @mock.patch("logprep.connector.dummy.output.DummyOutput.store_custom")
     @mock.patch("logprep.connector.dummy.input.DummyInput.get_next", return_value={"mock": "event"})
     def test_extra_dat_list_is_passed_to_store_custom(self, mock_get_next, mock_store_custom, _):
         self.pipeline._setup()
